Page should follow up to 3 redirects and ignore request or fragment params
The link passed to the page constructor may return a redirect. Need to load
the page in order to determine the true URL of the page. To avoid potentials
for redirect loops, only follow up to 3 redirects then stop.

Also the page should strip off any request or fragment portions of the URI;
these are significant for a web application but for a static website they
are not needed and can infinitely bloat the number of potential pages.